+++ packages/model-runtime/src/index.ts
@@ -79,4 +79,10 @@ export { consumeStreamUntilDone } from './utils/consumeStream';
 export { AgentRuntimeError } from './utils/createError';
 export { getModelPropertyWithFallback } from './utils/getFallbackModelProperty';
 export { getModelPricing } from './utils/getModelPricing';
+export {
+  applyModelExtendParams,
+  type ApplyModelExtendParamsContext,
+  type ModelExtendParams,
+  resolveDefaultThinkingLevelForModel,
+} from './utils/modelExtendParams';
 export { parseDataUri } from './utils/uriParser';
